Align metrics name. Remove duplicate classes
