+++ core/src/main/java/com/whylogs/core/metrics/ClassificationMetrics.java
@@ -1,30 +1,32 @@
 package com.whylogs.core.metrics;
 
+import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import com.whylogs.core.DatasetProfile;
 import com.whylogs.core.message.ScoreMatrixMessage;
 import com.whylogs.core.statistics.NumberTracker;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 
 @Slf4j
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Getter
 public class ClassificationMetrics {
   private List<String> labels;
+  private final String predictionField;
+  private final String targetField;
+  private final String scoreField;
   private NumberTracker[][] values;
 
-  public ClassificationMetrics() {
-    this(Lists.newArrayList(), newMatrix(0));
-  }
-
-  public static ClassificationMetrics of() {
-    return new ClassificationMetrics();
+  public ClassificationMetrics(String predictionField, String targetField, String scoreField) {
+    this(Lists.newArrayList(), predictionField, targetField, scoreField, newMatrix(0));
   }
 
   public List<String> getLabels() {
@@ -56,11 +58,28 @@ public class ClassificationMetrics {
     return res;
   }
 
-  public <T> void update(DatasetProfile datasetProfile, T prediction, T target, double score) {
+  public void track(Map<String, ?> columns) {
+    Preconditions.checkState(predictionField != null);
+    Preconditions.checkState(targetField != null);
+    val prediction = columns.get(predictionField);
+    val target = columns.get(targetField);
+    val scoreRaw = columns.get(scoreField);
+    double score = 0;
+    if (scoreRaw instanceof Number) {
+      score = ((Number) scoreRaw).doubleValue();
+    } else if (scoreRaw != null) {
+      try {
+        score = Double.parseDouble(scoreRaw.toString());
+      } catch (NumberFormatException e) {
+        log.warn("Failed to parse score: {}", scoreRaw, e);
+      }
+    }
+    this.update(prediction, target, score);
+  }
+
+  public <T> void update(T prediction, T target, double score) {
     val predictionText = textValue(prediction);
     val targetText = textValue(target);
-    datasetProfile.track("whylogs.metrics.predictions", predictionText);
-    datasetProfile.track("whylogs.metrics.targets", targetText);
 
     val x = labels.indexOf(predictionText);
     val y = labels.indexOf(targetText);
@@ -68,17 +87,19 @@ public class ClassificationMetrics {
       // happy case
       values[x][y].track(score);
     } else {
-      val newLabels = Lists.newArrayList(labels);
+      val newLabelsSet = Sets.newHashSet(labels);
 
       if (x < 0) {
-        newLabels.add(predictionText);
+        newLabelsSet.add(predictionText);
       }
-      if (y < 0 && !predictionText.equals(targetText)) {
-        newLabels.add(targetText);
+      if (y < 0) {
+        newLabelsSet.add(targetText);
       }
+
+      val newLabels = Lists.newArrayList(newLabelsSet);
       Collections.sort(newLabels);
 
-      final int newDim = newLabels.size();
+      final int newDim = newLabelsSet.size();
       val newValues = newMatrix(newDim);
 
       // first copy existing values to the new matrix
@@ -147,7 +168,8 @@ public class ClassificationMetrics {
     // copy the other object
     addMatrix(other.labels, other.values, newLabels, newValues);
 
-    return new ClassificationMetrics(newLabels, newValues);
+    return new ClassificationMetrics(
+        newLabels, targetField, predictionField, scoreField, newValues);
   }
 
   private void addMatrix(
@@ -175,7 +197,8 @@ public class ClassificationMetrics {
         copyValues[i][j] = copyValues[i][j].merge(values[i][j]);
       }
     }
-    return new ClassificationMetrics(Lists.newArrayList(this.labels), copyValues);
+    return new ClassificationMetrics(
+        Lists.newArrayList(this.labels), predictionField, targetField, scoreField, copyValues);
   }
 
   @NonNull
@@ -191,13 +214,27 @@ public class ClassificationMetrics {
       }
     }
 
+    builder.setPredictionField(predictionField);
+    builder.setTargetField(targetField);
+    builder.setScoreField(scoreField);
+
     return builder;
   }
 
   public static ClassificationMetrics fromProtobuf(ScoreMatrixMessage msg) {
+    if (msg == null || msg.getSerializedSize() == 0) {
+      return null;
+    }
+
     val labels = Lists.<String>newArrayList();
     for (int i = 0; i < msg.getLabelsCount(); i++) {
       labels.add(msg.getLabels(i));
+    }
+
+    if (msg.getLabelsCount() == 0 && msg.getScoresCount() > 0) {
+      // Not valid to have score without labels.
+      log.warn("Skipping classification ScoreMatrix: has scores but no labels");
+      return null;
     }
 
     final int n = labels.size();
@@ -208,6 +245,7 @@ public class ClassificationMetrics {
       values[row][col] = NumberTracker.fromProtobuf(msg.getScores(i));
     }
 
-    return new ClassificationMetrics(labels, values);
+    return new ClassificationMetrics(
+        labels, msg.getPredictionField(), msg.getTargetField(), msg.getScoreField(), values);
   }
 }
